Store index in package instead of on server
This file is tiny, so no point having it outsourced to the server.
If we add new catalogs to pyckles, we should release a new version anyway,
which can then also have an updated index file.

In preparation of using pooch for handling the downloads, we also need the
hashes of the fits files on the server. This index file seems like the
perfect place to store those as well, no need for a separate registry file
that would basically be the same list of files...