Use pooch to manage downloads and caching
Since the index is now stored in the package, no download function for the
index file aka cat list is required. Simply loading it when the package is
loaded should be fine, as the catalog list (index) will not change after
loading the package.

The use_cache argument is no longer required, as caching is managed by
pooch. Also, loading the cat list is now no longer a webtest.